Former-commit-id: 9e994262312c73c5c18c59e0984b6a70dceb644c

+++ src/main/scala/org/bitbucket/eunjeon/seunjeon/LNode.scala
@@ -50,7 +50,7 @@ object LNode {
 case class LNode(morpheme:Morpheme,
                  var startPos:Int,
                  var endPos:Int,
-                 var accumulatedCost:Int=Short.MaxValue) {
+                 var accumulatedCost:Int = Int.MaxValue) {
   var leftNode:LNode = null
   var isActive:Boolean=true
